Fixed compilation with MS VS2005 and its stupid "security" deprecations
git-svn-id: svn://svn.icculus.org/twilight/trunk/dpmaster@6500 d7cf8633-e32d-0410-b094-e92efae38249

+++ dpmaster.c
@@ -460,7 +460,7 @@ static qboolean SecureInit (void)
 
 	// Init the time and the random seed
 	crt_time = time (NULL);
-	srand (crt_time);
+	srand ((unsigned int)crt_time);
 
 	// Initialize the server list and hash table
 	if (!Sv_Init ())
